Modified value class serialization to use Converter
To make the JsonSerialize annotation assigned to the property work
